Refactor parser and allow it to return the local to the CLI app object rather than define it in an application-wide instance variable
